- moved home link out from Experiment menu(is it possible to go to home page clicking on the foodauthent logo) - add Misc menu in the nav - also the list of fingerprint ids are link

+++ foodauthent-osgi/foodauthent-ui-vuejs/src/components/model/Model.vue
@@ -23,7 +23,7 @@
                         </generalTable>
                     </b-tab>
                     <b-tab title="Create new">
-                        <generalForm :schema="schema" :model="model" :schemas="schemas" :options="formOptions" :save="save" :cancel="cancel" :pageType="pageType" :schemaIdHolder="schemaIdHolder" :showSuccess="showSuccess" :showError="showError"></generalForm>
+							<generalForm :schema="schema" :model="model" :schemas="schemas" :options="formOptions" :save="save" :cancel="cancel" :pageType="pageType" :schemaIdHolder="schemaIdHolder" :response="response" :showSuccess="showSuccess" :showError="showError" :loading="loading"></generalForm>		
                     </b-tab>
                 </b-tabs>
             </b-card>
@@ -86,7 +86,9 @@ export default {
             perPage: 10,
             filter: null,
             model: {},
+            response: "",
             pageType: "noType",
+            loading: false,
             schemas: schemas,
             itemsMetadata: {},
             resultsCount: 1,
@@ -133,13 +135,18 @@ export default {
             },
             save() {
                 let self = this;
+                self.loading = true;
                 console.log("POST BODY", JSON.stringify(this.model, undefined, 4));
                 saveModel(JSON.stringify(this.model, undefined, 4), self);
                 self.model = {}
+                self.showSuccess = false;
+                self.showError = false;
             },
             cancel() {
                 let self = this;
+                if(self.model["file-id"] != "" && self.model["file-id"] != null){
                 deleteFile(self.model["file-id"]);
+                }
                 self.model = {};
             },
             myRowClickHandler(record, index) {
